// resources/js/Request/HttpResponse.ts
export enum HttpResponse {
    TIMEOUT = "timeout",
    CANNOT_CONNECT = "cannot-connect",
    NOT_FOUND = "not-found",
    SERVER_ERROR = "server",
    UNAUTHORIZED = "unauthorized",
    FORBIDDEN = "forbidden",
    REJECTED = "rejected",
    UNKNOWN = "unknown",
    BAD_DATA = "bad-data",
    METHOD_NOT_ALLOWED = "method-not-allowed"
}